+++ entity/src/main/resources/db/migration/V1.19__update_template_doc_v8.sql
@@ -1,0 +1,73 @@
+-- ------------------------------------------------
+-- Version: v1.19
+--
+-- Description: Migration that updates the json document in the Template table for version 8
+-- -------------------------------------------------
+UPDATE Template set template_doc ='{
+  "name": "",
+  "assessment_type": {
+    "id": 1,
+    "name": "eosc pid policy"
+  },
+  "version": "",
+  "status": "",
+  "published": false,
+  "timestamp": "",
+  "actor": {
+    "id": 6,
+    "name": "PID Owner"
+  },
+  "organisation": {
+    "name": "",
+    "id": ""
+  },
+  "subject": {
+    "name": "",
+    "id": "",
+    "type": ""
+  },
+  "result": {
+    "compliance": null,
+    "ranking": null
+  },
+  "principles": [
+    {
+      "id": "P1",
+      "name": "Application",
+      "description": "PID application depends on unambiguous ownership, proper maintenance, and unambiguous identification of the entity being referenced.",
+      "criteria": [
+        {
+          "id": "C4",
+          "name": "Measurement",
+          "description": "The PID owner SHOULD maintain PID attributes.",
+          "imperative": "should",
+          "metric": {
+            "type": "number",
+            "result": null,
+            "value": null,
+            "benchmark": {
+              "equal_greater_than": 1
+            },
+            "algorithm": "sum",
+            "tests": [
+              {
+                "type": "binary",
+                "id": "T4",
+                "name": "Maintenance",
+                "description": "A test to determine if the entity (PID) attributes are being maintained.",
+                "text": "Do you regularly maintain the metadata for your object?",
+                "value": null,
+                "result":null,
+                "evidence_url": [],
+                "guidance": {
+                  "id": "G1",
+                  "description": "In practice, evaluation is very difficult, due to two factors: \\n - It requires that a sample of millions of PID owners be evaluated across all services, and \\n - Some entities may never have to be maintained and are, despite years of non-maintenance, up to date. \\n A measure of the mean update frequency of PIDs across a specific service, and monitoring its change over time against a benchmark, may be the only realistic assessment mechanism."
+                }
+              }
+            ]
+          }
+        }
+      ]
+    }
+  ]
+}' where id=1;